fixed repeating local storage and preference issues with Dani

+++ app.js
@@ -6,16 +6,35 @@ const loginButton = document.getElementById('login-button');
 const username = document.getElementById('username');
 const email = document.getElementById('email');
 
+
+
+
 loginButton.addEventListener('click', (event) => {
     event.preventDefault();
     if (email.value === '' || username.value === '') {
         return;
     }
     else {
-        const formData = new FormData(form);
-        const madeUser = makeUser(formData);
-        const stringed = JSON.stringify(madeUser);
-        localStorage.setItem('userData', stringed);
+        let userDataLocalStorage = localStorage.getItem('userData'); //find item in local storage called userData
+        if (!userDataLocalStorage){ // if it does not exist
+            const madeUser = new FormData(form); 
+            const userDataLocalStorage = makeUser(madeUser);
+            const stringed = JSON.stringify(userDataLocalStorage);
+            localStorage.setItem('userData', stringed);
+
+        } else {
+            userDataLocalStorage = JSON.parse(userDataLocalStorage); //if it does exist
+            const updatedUserInfo = new FormData(form);
+
+            userDataLocalStorage = updatedUserInfo; 
+        }
+        
+        
+        
+
+
+        // import user Data from local storage 
+
 
         window.location.replace ('./preferences/index.html');
     }
